Simplify Canvas Positioning
I simplified the positioning of the canvas in index.html so that it is automatically done, and there is no messing around with percentages or any of that.

Something that maybe we should maybe look into is automatic resizing of the canvas, since we're using explicit pixel values? Potential issues with it being too big or too small for certain screens.