Added flame particles to show how player is floating
I added the jetpack particle effect, to make it look like a jetpack is moving the player.

This can all be tweaked later, I just thought it was good to get it in earlier rather than later.

Currently the speed range is the same in x and y direction, and it's just kinda exploding out from the player's position; maybe in the playerAngle function we can figure out how to change this so it actually shoots out in the direction opposite the way the player is moving?